pyproject.toml
[tool.poetry]
name = "expensifain"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.1"
pypdf = "^4.2.0"
langchain-community = "^0.2.1"
litellm = "^1.38.8"
langfuse = ">=2.0.0"
diskcache = "^5.6.3"
ocrmac = "^0.1.6"
gradio = "^4.32.2"
ollama = "^0.2.1"


[tool.poetry.group.extras.dependencies]
langchain-google-genai = "^1.0.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
huggingface-hub = "^0.23.2"
pip = "^24.0"
fsspec = "2023.9.2"
pillow = "^10.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
